De-inline extract_number and move it back into cborparser.c
This function wasn't getting inlined anyway, might as well make it have
one canonical copy and make that get called.
